rm old dataset, rename geojson
