+++ app/components/Nav.tsx
@@ -15,7 +15,7 @@ export default function Nav() {
   }
 
   return (
-    <div className="bg-yellow-300 flex items-center py-2 px-2 gap-3">
+    <div className="bg-yellow-300 flex items-center py-2 px-8 gap-3">
       <span className="flex flex-1 gap-3">
         <a className="flex" href="/">
           <img
